Move middlewares to interface layer
Things related to HTTP must be on the interface layer
